# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  university.aulaRequest:
    properties:
      color_idx:
        type: integer
      day_of_week:
        type: integer
      room:
        type: string
      semester:
        type: string
      subject_name:
        type: string
      time_start:
        type: string
    type: object
...
  university.materiaRequest:
    properties:
      absences:
        type: integer
      attendance:
        type: number
      color_idx:
        type: integer
      credits:
        type: integer
      end_date:
        type: string
      grade:
        type: number
      icon:
        type: string
      max_absences:
        type: integer
      name:
        type: string
      professor:
        type: string
      semester:
        type: string
      start_date:
        type: string
      subject_status:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: 'API do BeePlanner — autenticação via JWT (Bearer). Todas as rotas
    /api/* (exceto /auth) exigem o header Authorization: Bearer <token>.'
  title: BeePlanner API
  version: "1.0"
paths:
...
  /university/schedule:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar grade horária
      tags:
      - University
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da aula
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/university.aulaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar aula na grade
      tags:
      - University
  /university/schedule/{id}:
    delete:
      parameters:
      - description: ID da aula
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar aula da grade
      tags:
      - University
  /university/subjects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar matérias
      tags:
      - University
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da matéria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/university.materiaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar matéria
      tags:
      - University
  /university/subjects/{id}:
    delete:
      parameters:
      - description: ID da matéria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar matéria
      tags:
      - University
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da matéria
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/university.materiaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Editar matéria
      tags:
      - University
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <seu_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"